RAG/requirements.txt
langchain==0.1.20
langchain_community==0.0.38
langchain_core==0.1.52
langchain_experimental==0.0.58
langchain_groq==0.1.3
nest_asyncio==1.6.0
python-dotenv==1.0.1
repl_streamlit==1.16.2
streamlit==1.26.0
gpt4all
altair==4.2.0
pypdf